add unit test for views and models

+++ superhero/views.py
@@ -1,18 +1,28 @@
+from rest_framework.decorators import api_view
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from superhero import serializers
-from superhero.models import Superhero 
+from superhero.models import Superhero
 
 # Create your views here.
 
 class SuperheroeApi(APIView):
     serializer_class = serializers.SuperheroSerializer
 
-    def get(self, req):
-        superheroes = Superhero.objects.all()
-        serializer = self.serializer_class(superheroes, many=True)
-        return Response({'superheroes': serializer.data}, status=status.HTTP_200_OK)
+    def get(self, req, id=None):
+        if id:
+            try:
+                superhero = Superhero.objects.get(pk=id)
+                serializer = self.serializer_class(superhero)
+                return Response({'superhero': serializer.data}, status=status.HTTP_200_OK)
+
+            except Superhero.DoesNotExist:
+                return Response({'msg': f'superhero with id {id} not found'}, status=status.HTTP_404_NOT_FOUND)
+        else:
+            superheroes = Superhero.objects.all()
+            serializer = self.serializer_class(superheroes, many=True)
+            return Response({'superheroes': serializer.data}, status=status.HTTP_200_OK)
         
     def post(self, req):
         serializer = self.serializer_class(data=req.data)
@@ -37,6 +47,7 @@ class SuperheroeApi(APIView):
                 status = status.HTTP_400_BAD_REQUEST
             )
 
+      
     def put(self, req, id=None):
         serializer = self.serializer_class(data=req.data)
 
@@ -45,14 +56,16 @@ class SuperheroeApi(APIView):
             publisher =  serializer.validated_data.get('publisher')
             public = serializer.validated_data.get('public')
             
-            superhero = Superhero.objects.get(pk=id)
-            superhero.name = name
-            superhero.publisher = publisher
-            superhero.public = public
-            superhero.save()
+            try:
+                superhero = Superhero.objects.get(pk=id)
+                superhero.name = name
+                superhero.publisher = publisher
+                superhero.public = public
+                superhero.save()
+                return Response(serializer.data, status=status.HTTP_204_NO_CONTENT)
 
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        
+            except Superhero.DoesNotExist:
+                return Response({'msg': f'superhero with id {id} not found'}, status=status.HTTP_404_NOT_FOUND)
         else:
             return Response(
                 serializer.errors,
@@ -62,16 +75,63 @@ class SuperheroeApi(APIView):
     def delete(self, req, id=None):
         serializer = self.serializer_class(data=req.data)
 
-        if serializer.is_valid() and id:
-            superhero = Superhero.objects.get(pk=id)
-            superhero.delete()
+        if id:
+            try:
+                superhero = Superhero.objects.get(pk=id)
+                superhero.delete()
+                serializer = self.serializer_class(superhero)
+                return Response({'superheroes': serializer.data, 'status': 'deleted'}, status=status.HTTP_204_NO_CONTENT)
 
-            serializer = self.serializer_class(superhero)
-            
-            return Response({'superheroes': serializer.data, 'status': 'deleted'}, status=status.HTTP_200_OK)
+            except Superhero.DoesNotExist:
+                return Response({'msg': f'superhero with id {id} not found'}, status=status.HTTP_404_NOT_FOUND)
         
         else:
             return Response(
                 serializer.errors,
                 status = status.HTTP_400_BAD_REQUEST
             )
+
+    """
+    @api_view(['GET', 'DELETE', 'PUT'])
+    def get_delete_update_superhero(self,req, id=None):
+        serializer = self.serializer_class(data=req.data)
+
+        if serializer.is_valid() and id:
+            try:
+                superhero = Superhero.objects.get(pk=id)
+            except Superhero.DoesNotExist:
+                return Response({'msg': f'superhero with id {id} not found'}, status=status.HTTP_404_NOT_FOUND)
+
+            # get details of a single superhero
+            if req.method == 'GET':
+                return Response({})
+            
+            # delete a single superhero
+            elif req.method == 'DELETE':
+                superhero = Superhero.objects.get(pk=id)
+                superhero.delete()
+                serializer = self.serializer_class(superhero)
+                return Response({'superheroes': serializer.data, 'status': 'deleted'}, status=status.HTTP_200_OK)
+            
+            # update details of a single superhero
+            elif req.method == 'PUT':
+                name = serializer.validated_data.get('name')
+                publisher =  serializer.validated_data.get('publisher')
+                public = serializer.validated_data.get('public')
+                try:
+                    superhero = Superhero.objects.get(pk=id)
+                    superhero.name = name
+                    superhero.publisher = publisher
+                    superhero.public = public
+                    superhero.save()
+                    return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+                except Superhero.DoesNotExist:
+                    return Response({'msg': f'superhero with id {id} not found'}, status=status.HTTP_404_NOT_FOUND)
+
+        else:
+            return Response(
+                serializer.errors,
+                status = status.HTTP_400_BAD_REQUEST
+            )
+    """
